replace 'docopt' with 'structopt' in cargo-rm
